move test methods to utils, add dummy entry for populate test
